+++ resources/views/default/_partials/header.blade.php
@@ -14,6 +14,7 @@
 		</ul>
 
 		<ul class="nav navbar-nav navbar-right">
+			<li><a href="{{ route('admin.env.editor') }}"><i class="fa fa-btn fa-gear"></i> ENV Editor</a></li>
 			@stack('navbar.right')
 		</ul>
 	</div>